Create setup script
Setting up my workstation gets a bit tedious when moving to a new PC so
this is my effort to save having to remember all the dependencies to get
my system working just the way I like it.

Walk through the setup a bit to stop needing to repeat sections on
reruns of the script.