Create script to build latest Neovim
Debian is my distro of choice and Neovim is not available in Debian
Stable yet. In the meantime the simplist way to get the benefits of
Neovim is to build it from source.

Produced a script to automate cloning and building the latest release of
Neovim. Also include an uninstall script just in case.