Fix #1. Expand language support for YouCompleteMe
YouCompleteMe was not set to include full language support in Vim
causing an error when attempting to use autocompletion for some file
types.

Debian does not yet package Rust so out of the box support in
YouCompleteMe has been omitted. If this is needed Rust should be
manually installed. https://www.rust-lang.org/en-US/install.html