Remove use of 'nmap' in favour of 'noremap'
Apparently if 'nmap' is used but the shortcut is already bound to
another mapping vim can freeze up because of the mapping collision.

Using 'noremap' will apply the same mappings but won't cause vim to
freeze up it'll just used whatever the last remapping was.

See: https://dev.to/mr_destructive/vim-keymapping-guide-3olb